- Refactored all anonymous request functions into a function file
