+++ index.js
@@ -4,6 +4,7 @@ const app = express();
 const methodOverride = require('method-override')
 const reactEngine = require('express-react-views').createEngine();
 const moment = require('moment')
+const callback = require('./functions')
 
 //JSON
 const file = 'recipes.json';
@@ -29,235 +30,16 @@ app.use(express.static("."));
 // ******************************************
 // ******************************************
 
-app.get('/recipes/new', (request, response) => {
-  //render new recipe form
-  jsonfile.readFile(file, (err, obj)=>{
+//CRUD requests
+app.get('/recipes/new', callback.newRecipe)
+app.post('/recipes', callback.saveNewRecipe)
+app.get('/recipes/:id', callback.showRecipe)
+app.get('/recipes/:id/edit', callback.editRecipe)
+app.put('/recipes/:id', callback.updateRecipe)
+app.get('/recipes/:id/delete', callback.deleteDisplay)
+app.delete('/recipes/:id', callback.deleteRecipe)
+app.get('/recipes', callback.showAll)
+app.get('/', callback.home)
 
-    const data = {
-      recipes: obj.recipes
-    }
-
-    response.render('new', data)
-  })
-})
-
-app.post('/recipes', (request, response) => {
-  //get newRecipe from post request form
-  const newRecipe = request.body
-
-  jsonfile.readFile(file, (err, obj) => {
-    if (err) {
-      return console.log(err)
-    }
-    newRecipe.created = "Created on " + moment().format('MMMM Do YYYY, h:mm:ss a')
-    newRecipe.id = obj.recipes.length + 1
-
-    //push recipe into json file
-    obj.recipes.push(newRecipe)
-    //retrieve pushed recipe's ID
-    const newRecipeID = obj.recipes.length - 1
-    //create data obj with the new item, for rendering success message
-    const data = {
-      recipes: obj.recipes,
-      recipe: obj.recipes[newRecipeID],
-      success: "Item added successfully!"
-    }
-    jsonfile.writeFile(file, obj, (err) => {
-      if (err) {
-        return console.log(err)
-      }
-      response.render('displayid', data)
-      // response.send("Success!")
-    })
-    //END readFile
-  })
-
-  //END app.post
-})
-
-app.get('/recipes/:id', (request, response) => {
-  //get id
-  const id = request.params.id
-
-  jsonfile.readFile(file, (err, obj) => {
-    if (err) {
-      return console.log(err)
-    }
-
-    //get recipe object by id
-    const requestedRecipe = obj.recipes[id]
-
-    //store recipe object in data object
-    const data = {
-      //pass whole recipes array for navbar
-      recipes: obj.recipes,
-      //pass recipe for the requested ID
-      recipe: requestedRecipe
-    }
-
-    //render display .jsx and pass recipe as data
-    response.render('displayid', data)
-    //END readFile
-  })
-  //END app.get
-})
-
-app.get('/recipes/:id/edit', (request, response) => {
-  //get id
-  const id = request.params.id
-
-  jsonfile.readFile(file, (err, obj) => {
-    if (err) {
-      return console.log(err)
-    }
-    //requestedObj content set to object in recipe book
-    const requestedRecipe = obj.recipes[id]
-    //create data object to pass to edit form
-    const data = {
-      recipes: obj.recipes,
-      id: id,
-      recipe: requestedRecipe
-    }
-
-    response.render('edit', data)
-    //END readFile
-  })
-  //END app.get
-})
-
-app.put('/recipes/:id', (request, response) => {
-  //get id
-  const id = request.params.id
-  //store submitted change 
-  const editedRecipe = request.body
-
-  jsonfile.readFile(file, (err, obj) => {
-    if (err) {
-      return console.log(err)
-    }
-
-    //copy editedRecipe into recipe object
-    Object.assign(obj.recipes[id], editedRecipe)
-    //create data object for rendering success message
-    const data = {
-      recipes: obj.recipes,
-      success: "Edited successfully!",
-      recipe: editedRecipe
-    }
-
-    jsonfile.writeFile(file, obj, (err) => {
-      if (err) {
-        return console.log(err)
-      }
-      response.render('displayid', data)
-    })
-    //END readFile
-  })
-  //END app.put
-})
-
-app.get('/recipes/:id/delete', (request, response) => {
-  //get id
-  const id = request.params.id
-
-  jsonfile.readFile(file, (err, obj) => {
-    if (err) {
-      return console.log(err)
-    }
-    //requestedObj content set to object in recipe book
-    const requestedRecipe = obj.recipes[id]
-    //create data object to pass to delete form
-    const data = {
-      recipes: obj.recipes,
-      id: id,
-      recipe: requestedRecipe
-    }
-
-    response.render('delete', data)
-    //END readFile
-  })
-  //END app.get
-})
-
-app.delete('/recipes/:id', (request, response) => {
-  // get id
-  const id = request.params.id
-
-  jsonfile.readFile(file, (err, obj) => {
-    if (err) {
-      return console.log(err)
-    }
-
-    //remove Object at id
-    obj.recipes.splice(id, 1)
-
-    let newID = 1
-    for (let recipe of obj.recipes){
-      recipe.id = newID
-      newID++
-    }
-
-
-
-    jsonfile.writeFile(file, obj, (err) => {
-      if (err) {
-        return console.log(err)
-      }
-
-      const data = {
-        recipes: obj.recipes
-      }
-
-      response.render('home', data)
-    })
-    //END readFile
-  })
-  //END app.delete
-})
-
-app.get('/recipes', (request,response)=>{
-  
-  jsonfile.readFile(file, (err, obj)=>{
-    const data = {
-      recipes: obj.recipes
-    }
-
-    response.render('overview', data)
-  })
-})
-
-app.get('/', (request, response)=>{
-
-  jsonfile.readFile(file, (err, obj)=>{
-
-    response.render('home', obj)
-  })
-})
-
-//NEXT FEATURE
-// app.get('/recipes/sort', (request, response)=>{
-
-//   jsonfile.readFile(file, (err, obj)=>{
-
-//     obj.recipes.sort(
-//       function (a, b) {
-//         if (a.id > b.id) {
-//           return -1;
-//         } else if (a.id < b.id) {
-//           return 1;
-//         } else {
-//           return 0;
-//         }
-//       }
-//     )
-
-//     const data = {
-//       recipes: obj.recipes
-//     }
-
-//       response.render('home', data)
-//   })
-  
-// })
-
+//Listen for requests
 app.listen(3000, () => console.log("Listening.....TO YOUR INPUT"))
