+++ views/overview.jsx
@@ -21,7 +21,7 @@ render() {
         )
     })
 
-    
+   
 
 return (
